% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lake_info.R
\name{get_lake_info}
\alias{get_lake_info}
\title{Get a summary of available information about a lake or lakes from the
UK CEH Lakes Portal}
\usage{
get_lake_info(...)
}
\arguments{
\item{...}{A series of lake id numbers to searched. Can use ranges etc.
#'}
}
\value{
A data frame containing the available information about the specified
lake or lakes. The 'Biology' information on the Lakes Portal webpages is not
included, but all other information is provided.
}
\description{
Produces a summary of available information about a lake or
lakes from the UK CEH Lakes Portal by responsibly webscraping lake
information pages.
#
}
\examples{
# get information for Loch Katrine and Loch Lomond
get_lake_info (24447, 24531)

# can also
}
