% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_buffer.R
\name{search_buffer}
\alias{search_buffer}
\title{Search database of for lakes within a set buffer distance of a point}
\usage{
search_buffer(xcoord, ycoord, buffer_distance)
}
\arguments{
\item{xcoord}{The x coordinate, in British National Grid, of the point to
base the buffer on.}

\item{ycoord}{The y coordinate, in British National Grid, of the point to
base the buffer on.}

\item{buffer_distance}{The buffer distance (radius) to be searched around
the point in metres.}
}
\value{
A data frame containing the details (Name, wbid, Country) of all
the lakes that are within the specified buffer distance of the point as well
as the distance (in metres) from each lake to the point.
}
\description{
Searches the listing of UK CEH Lakes Portal lake sites to find
lakes that are within a specified buffer distance from the point given.

Coordinates are specified as British National Grid X and Y and are checked
to ensure that they are within the required ranges.
}
\examples{
# search for sites within 1km of the point 315231, 642134
search_buffer(335792, 501379, 2000)

}
