% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_names.R
\name{search_names}
\alias{search_names}
\title{Search database of site names}
\usage{
search_names(string = NULL)
}
\arguments{
\item{string}{The search string to be matched (case-sensitive). Will match
whole or partial strings in the column values.
#'}
}
\value{
A data frame containing the details (Name & lakeid) of all the sites
that match the search string (full or partial matches).
}
\description{
Searches the listing of UK CEH Lakes Portal sites to find names
that contain the string provided.

The search is done on an internal copy of the UK lakes listing rather than
connecting to the UKCEH Lakes Portal site.
}
\examples{
# search for sites containing "Tarn" in the name
search_names(string="Tarn")

}
