% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uklakes-package.R
\docType{package}
\name{uklakes-package}
\alias{uklakes}
\alias{uklakes-package}
\title{uklakes: Retrieves Lake Information from UK CEH Lakes Portal}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Retrieves a summary of the lake information from the UK CEH Lake Portal for a lake or series of lakes, based on the lakeid. Can also search for lakeid values based on strings in the lake name. For details of the information available, see the Lakes Portal Website (\url{https://uklakes.ceh.ac.uk/}.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://robbriers.github.io/uklakes/}
}

}
\author{
\strong{Maintainer}: Rob Briers \email{r.briers@napier.ac.uk} (\href{https://orcid.org/0000-0003-0341-1203}{ORCID})

}
\keyword{internal}
